pyproject.toml
[project]
name = "sdv_mcp_demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"fastapi>=0.115.12",
	"llama-index>=0.12.31",
	"llama-index-embeddings-huggingface>=0.5.3",
	"llama-index-llms-openai-like>=0.3.4",
	"llama-index-llms-siliconflow>=0.2.1",
	"llama-index-llms-deepseek",
	"llama-index-vector-stores-postgres>=0.4.2",
	"mcp-server>=0.1.4",
	"mysql-connector-python>=9.3.0",
	"numpy<2.0",
	"paho-mqtt>=2.1.0",
	"pandas>=2.2.3",
	"python-dotenv>=1.1.0",
	"ruff>=0.11.4",
	"sse-starlette>=2.2.1",
	"uvicorn>=0.34.1",
	"torch==2.2.2",
	"mcp[cli]>=1.7.1",
	"httpx>=0.28.1",
	"llama-index-tools-mcp>=0.1.2",
]
